requirements.txt
psl-dns
nextcord
python-dotenv
onami
aiohttp
psycopg2-binary
nextcord-ext-help-commands
nextcord-ext-menus
python-whois
forex-python
